Fixed inserting new employee into database

+++ src/main/java/com/alaythiaproductions/customspringbootmvc/repositories/SkillRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SkillRepository extends JpaRepository<Skill, Long> {
+public interface SkillRepository extends JpaRepository<Skill, Integer> {
 
     List<Skill> findByEmployee(Employee employee);
 }
